-Add DbgEng service GUIDs -Add WinDbgExtensionAPI.QueryTargetInterface -Fix PDB1 delegates not using ThisCall, which is required on x86 -Add ITypeLib As<T> extension methods -Don't re-wrap interface methods when using As<T> if the wrapper is already of the desired target type -Rename DllGetClassObjectInterfaces.ClassFactory() to GetClassFactory() -Fix CorDebugGenericValue.Value incorrectly being a property instead of a method -Fix DbgEng wrappers not releasing sub-interfaces on DIspose -Fix several parameters with incorrect enum/token types -Add TTD LiveRecorder wrapper -Change DEBUG_STATUS to an enum -Use an ushort on PDB1 symbol reclen fields -Fix PDB1 symbol packing

+++ ClrDebug/Managed/Metadata/MetaDataImport.cs
@@ -2914,7 +2914,7 @@ namespace ClrDebug
             [Out] out CorParamAttr pdwAttr,
             [Out] out CorElementType pdwCPlusTypeFlag,
             [Out] out IntPtr ppValue,
-            [Out] out IntPtr pcchValue);*/
+            [Out] out int pcchValue);*/
             mdMethodDef pmd;
             int pulSequence;
             char[] szName;
@@ -2923,7 +2923,7 @@ namespace ClrDebug
             CorParamAttr pdwAttr;
             CorElementType pdwCPlusTypeFlag;
             IntPtr ppValue;
-            IntPtr pcchValue;
+            int pcchValue;
             HRESULT hr = Raw.GetParamProps(tk, out pmd, out pulSequence, null, cchName, out pchName, out pdwAttr, out pdwCPlusTypeFlag, out ppValue, out pcchValue);
 
             if (hr != HRESULT.S_FALSE && hr != HRESULT.ERROR_INSUFFICIENT_BUFFER && hr != HRESULT.S_OK)
